rbeads - R implementation of Bias Elimination Algorithm for Deep Sequencing

BEADS is a normalization scheme that corrects nucleotide composition bias, mappability variations and differential local DNA structural effects in deep sequencing data. In high-throughput sequencing data, the recovery of sequenced DNA fragments is not uniform along the genome. In particular, GC-rich sequences are often over-represented and AT-rich sequences under-represented in sequencing data. In addition, the read mapping procedure also generates regional bias. Sequence reads that can be mapped to multiple sites in the genome are usually discarded. Genomic regions with high degeneracy therefore show lower mapped read coverage than unique portions of the genome. Mappability varies along the genome and thus creates systematic bias. Furthermore, local DNA or chromatin structural effects can lead to coverage inhomogeneity of sequencing data.

Details

For detailed instruction how to run BEADS see beads function documentation.

References

http://beads.sourceforge.net/ http://www.ncbi.nlm.nih.gov/pubmed/21646344

See also

beads sumBAMinputs

Examples

# Get the paths of example files
sample_bam <- system.file("extdata", "GSM1208360_chrI_100Kb_q5_sample.bam", package="rbeads")
input_bam <- system.file("extdata", "Input_fE3_AA169.bam", package="rbeads")
SummedInput_bw <- system.file("extdata", "Ce10_HiSeqFRMInput_UNIQ_bin25bp_chrI_100Kb_sample.bw", package="rbeads")
map_bw <- system.file("extdata", "ce10_mappability_chrI_100Kb_sample.bw", package="rbeads")
ref_fa <- system.file("extdata", "ce10_chrI_100Kb_sample.fa", package="rbeads")

# Set the directory where the output files will be crated
setwd(tempdir())

# Run BEADS for BAM input file
beads(sample_bam, input_bam, map_bw, ref_fa)

#> Importing refference genome...

#> Importing mappability [binning]...

#> Importing input...

#> Reading alignment file [/Users/przemol/code/rbeads/inst/extdata/Input_fE3_AA169.bam]...	<0.025s>
#> Processing alignment file [uniq=TRUE]:...	<0.067s>

#> INFO: 22409 out of 27847 (80.47%) reads mapped uniquely with 10 quality cutoff: [All=27847, Mapped=27847, Qmapped=27807, Quniq=22409].

#> Calculate input GCcontent....	<1.062s>
#> Sample genome for GCcontent....	<0.006s>
#> Calculate histograms for genomic (a) nad and sample (b) GC content...

#> 	<0.051s>
#> Calculate GC weighting vector...

#> 	<0.266s>
#> Assign GC score to every read...	<0.001s>
#> Prepare GCscore-enriched GRanges...	<0.029s>
#> Calculate GC weighted coverage...	<0.023s>
#> Mappability correction...	<0.016s>

#> 	INFO: Total 3% of genome in masked.

#> Reading alignment file [/Users/przemol/code/rbeads/inst/extdata/GSM1208360_chrI_100Kb_q5_sample.bam]...	<0.016s>
#> Processing alignment file [uniq=TRUE]:...	<0.04s>

#> INFO: 8216 out of 10373 (79.21%) reads mapped uniquely with 10 quality cutoff: [All=10373, Mapped=10373, Qmapped=10346, Quniq=8216].

#> Calculate coverage...	<0.013s>

#> INFO: a = 18

#> Calling enriched regions...	<0.036s>
#> Mask out reads in enriched regions...	<0.011s>

#>  INFO: percentage of reads in non-enriched regions: 57.17%

#> Calculate input GCcontent....	<0.151s>
#> Sample genome for GCcontent....	<0.019s>
#> Calculate histograms for genomic (a) nad and sample (b) GC content...

#> 	<0.034s>
#> Calculate GC weighting vector...

#> 	<0.016s>
#> Assign GC score to every read...	<0s>
#> Prepare GCscore-enriched GRanges...	<0.004s>
#> Calculate GC weighted coverage...	<0.012s>
#> Mappability correction...	<0.01s>

#> 	INFO: Total 3% of genome in masked.

#> DivStep: Divideing...	<0.467s>
#> DivStep [median ]: Scaling...	<0.028s>

#> INFO: BEADS score scaling coefficient = 0.297

#> Exporing ER...

#> Exporing BigWig tracks...

#> Exporting to BigWiggle: GSM1208360_chrI_100Kb_q5_sample_BEADS_linear_1bp.bw

# Run BEADS for SummedInput (BigWig) input file
beads(sample_bam, SummedInput_bw, map_bw, ref_fa)

#> Importing refference genome...

#> Importing mappability [binning]...

#> Importing input...

#> Reading alignment file [/Users/przemol/code/rbeads/inst/extdata/GSM1208360_chrI_100Kb_q5_sample.bam]...	<0.014s>
#> Processing alignment file [uniq=TRUE]:...	<0.033s>

#> INFO: 8216 out of 10373 (79.21%) reads mapped uniquely with 10 quality cutoff: [All=10373, Mapped=10373, Qmapped=10346, Quniq=8216].

#> Calculate coverage...	<0.011s>

#> INFO: a = 18

#> Calling enriched regions...	<0.023s>
#> Mask out reads in enriched regions...	<0.008s>

#>  INFO: percentage of reads in non-enriched regions: 57.17%

#> Calculate input GCcontent....	<0.135s>
#> Sample genome for GCcontent....	<0.015s>
#> Calculate histograms for genomic (a) nad and sample (b) GC content...

#> 	<0.032s>
#> Calculate GC weighting vector...

#> 	<0.017s>
#> Assign GC score to every read...	<0s>
#> Prepare GCscore-enriched GRanges...	<0.004s>
#> Calculate GC weighted coverage...	<0.013s>
#> Mappability correction...	<0.012s>

#> 	INFO: Total 3% of genome in masked.

#> DivStep: Divideing...	<0.01s>
#> DivStep [median ]: Scaling...	<0.015s>

#> INFO: BEADS score scaling coefficient = 0.081

#> Exporing ER...

#> Exporing BigWig tracks...

#> Exporting to BigWiggle: GSM1208360_chrI_100Kb_q5_sample_BEADS_linear_1bp.bw